maintainers-list: create a field `inactivityReason`
Default null; should be a free-form string explaining why the maintainer
became inactive.

The most immediate use of this attribute is to mark maintainers that are
not contributing to Nixpkgs from a long amount of time. Nonetheless it can
be used to mark any *long term inactivity*, including but not limited to:

- Sabbatical leave
- Retirement
- Intimate issues
- Force majeure

In principle, a third person can set this attribute. For the sake of a good
etiquette, in such case at least one contact attempt should be issued and
carried out before effectively committing it to the codebase.

This attribute can be employed to many useful activities, including but not
limited to:

- Treewide automation
  - Removal of inactive maintainers from packages
  - Orphaning alerts
- Filter automatic notifications