duckstation: detach sources acquisition to sources.nix
The build scripts used by upstream require Git commands. Because of this, a
`deepClone` is required by fetchFromGitHub.

However, keeping `.git` directory around is detrimental to determinism.

Because of this, we need to run the Git commands "manually", save their outputs
and finally remove `.git`.

In order to not pollute the `package.nix`, this peculiar source acquisition is
detached to its own file and `callPackage`d.

Before this Jesuitic hack, the aforementioned data were obtained and filled by
human hands...